time indicator moved to DebugInfo
