+++ src/component/DebugInfo.module.styl
@@ -1,3 +1,8 @@
 .DebugInfo
   background-color black
-  opacity .2
+  opacity .4
+
+  font-size 1.8em
+
+  h1
+    font-size 1.2em
