+++ src/component/River.tsx
@@ -1,14 +1,14 @@
 import ChatComment from "@/type/ChatComment";
 import {
+  Accessor,
   Component,
-  createEffect,
   createResource,
   createSignal,
   For, 
 } from "solid-js";
 
 import {
-  getNewChatMessageObserver, getVideo, 
+  getNewChatMessageObserver, 
 } from "../script/InPageElement";
 
 import Flow from "./Flow";
@@ -20,7 +20,9 @@ type ScheduledToFlow<T> = {
   duplicatedCount: number,
 }
 
-const River: Component = () => {
+const River: Component<{
+  playbackTime: Accessor<number>
+}> = (props) => {
   const [comments, setComments] = createSignal<ScheduledToFlow<ChatComment>[]>([]);
   const [inInitDraws, setInInitDraws] = createSignal<{
     id: string,
@@ -35,9 +37,6 @@ const River: Component = () => {
     ]);
   const removeInInitDraw = (id: string) =>
     setInInitDraws((prev) => prev.filter((it) => it.id != id)); 
-  const [videoResource] =
-    createResource(() => getVideo());
-  const [time, setTime] = createSignal(0);
   createResource(() =>
     getNewChatMessageObserver((added) => {
       const initDrawCount =
@@ -47,36 +46,24 @@ const River: Component = () => {
           
       setComments((prev) => {
         const timeStumped =
-            added.map((it) => ({
-              value: it,
-              startTime: time(),
-              inInitDraw: false,
-              duplicatedCount: initDrawCount,
-            }));
+          added.map((it) => ({
+            value: it,
+            startTime: props.playbackTime(),
+            inInitDraw: false,
+            duplicatedCount: initDrawCount,
+          }));
         return [...prev, ...timeStumped];
       });
     })
   );
 
-  createEffect(() => {
-    const video = videoResource();
-    if (!video) return;
-    const loop = () => {
-      setTime(video.currentTime);
-      window.requestAnimationFrame(loop);
-    };
-    window.requestAnimationFrame(loop);
-  });
-  
   let element!: HTMLDivElement;
   return (
     <div ref={element} class={styles.River}>
-      <h1>River</h1>
-      <p>time: {time()}</p>
       <For each={comments()}>{(it) => (
         <Flow
           screenLength={element.clientWidth}
-          currentTime={time() - it.startTime}
+          currentTime={props.playbackTime() - it.startTime}
           duplicatedCount={it.duplicatedCount}
           comment={it.value}
           setInInitDraw={setInInitDraw}
